Fixed an int overflow bug

+++ include/solver.h
@@ -19,7 +19,7 @@ typedef struct {
 	int		ngrids;   // num of grids in this level
 	int		*gridId; // Grid Id of each grid in a given level
 	double		(*h)[2]; // Delta h in reference domain // ! Remove
-	int		(*ranges)[2]; // ranges of global indices for each grid
+	long int	(*ranges)[2]; // ranges of global indices for each grid
 	ArrayInt2d	global;  // global to grid map// ! Remove
 	ArrayInt2d	*grid;   // grid to global map// ! Remove
 } Level;
